Factored cfg,data, and cli py files from 3 modules
