+++ src/cli/__init__.py
@@ -1,0 +1,4 @@
+""" Command line interface for the package. """
+from .cli import custom_arg_parser
+
+__all__ = ["custom_arg_parser"]
