+++ src/data/__init__.py
@@ -1,0 +1,4 @@
+""" Data module interface."""
+from .dataset import create_dataset
+
+__all__ = ["create_dataset"]
